Add wasm opengl implementation.
This commit lays the groundwork of implementing
the core rendering engine along with a wasm/opengl
backend.

This commit also includes assimp and glm sources.
This is to allow these 3rd party libraries to be
compiled to wasm. It may be worth to trimming
these dependencies down to their bare essentials.

This commit also adds a dependency on SDL2, which
needs to be installed when building.

It is not assumed that the current state of the
engine is working. Bugs are expected.